+++ AngularWebApp/src/app/components/menu/menu.component.ts
@@ -1,7 +1,9 @@
 import { Component } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Responses } from '../../models/responses.model';
 import { BackendApiService } from '../../services/backend-api.service';
+import { DetailDialogComponent } from '../detail-dialog/detail-dialog.component';
 
 @Component({
   selector: 'app-menu',
@@ -14,7 +16,8 @@ export class MenuComponent {
 
   constructor(private route: Router,
     private route_: ActivatedRoute,
-    private backendApiService: BackendApiService) { }
+    private backendApiService: BackendApiService,
+    public dialog: MatDialog) { }
 
   username = String(this.route_.snapshot.paramMap.get('username'));
 
@@ -29,15 +32,27 @@ export class MenuComponent {
   }
 
   onClickDeleteUser() {
-    this.backendApiService.deleteUser(this.username).subscribe(
-      data => {
-        this.response = data;
+    const mdConfig = new MatDialogConfig();
+    mdConfig.disableClose = true;
+    mdConfig.width = '400px';
+    const dialogRef = this.dialog.open(DetailDialogComponent, mdConfig);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.backendApiService.deleteUser(this.username).subscribe(
+          data => {
+            this.response = data;
+          }
+        )
+        this.route.navigateByUrl('/home');
       }
-    )
-    this.route.navigateByUrl('/home');
+    })
+
+ 
   }
 
   onClickExit() {
     this.route.navigateByUrl('/home');
   }
+
 }
